requirements.txt
anyio==3.7.1
asgiref==3.7.2
asyncpg==0.25.0
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.2
dnspython==2.5.0
ecdsa==0.18.0
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.75.2
greenlet==3.0.3
h11==0.14.0
idna==3.6
passlib==1.7.4
psycopg2-binary==2.9.3
pyasn1==0.5.1
pycparser==2.21
pydantic==1.10.14
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.36
starlette==0.17.1
typing_extensions==4.9.0
uvicorn==0.17.6
